gpt_server/requirements.txt
flask
flask-cors
pymysql
python-dotenv
flask-jwt-extended